requirements.txt
-e git+git://github.com/wuurrd/python-yr.git#egg=python-yr
python-dateutil
